Refactor: DI를 각 모듈로 이동 및 Lint 수행
