+++ data/expense/src/main/java/com/kappzzang/jeongsan/repositoryimpl/ExpenseFakeRepositoryImpl.kt
@@ -5,10 +5,10 @@ import com.kappzzang.jeongsan.model.ExpenseItem
 import com.kappzzang.jeongsan.model.ExpenseListResponse
 import com.kappzzang.jeongsan.model.ExpenseState
 import com.kappzzang.jeongsan.repository.ExpenseRepository
+import java.util.Date
 import javax.inject.Inject
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
-import java.util.Date
 
 data class ExpenseListCachingKey(val expenseState: ExpenseState, val groupId: String)
 
@@ -47,8 +47,8 @@ class ExpenseFakeRepositoryImpl @Inject constructor(
         price = 15800,
         // 임시 지출 이미지 주소 (카카오테크 캠퍼스)
         expenseImageUrl = "https://www.kakaotechcampus.com/fileUpDownload/" +
-                "download.do?p_savefile=gatepage_20230330053504999_1.png&p_realfile=" +
-                "GNB+%EB%A1%9C%EA%B3%A0%28%EB%B3%B4%EB%9D%BC%29.png",
+            "download.do?p_savefile=gatepage_20230330053504999_1.png&p_realfile=" +
+            "GNB+%EB%A1%9C%EA%B3%A0%28%EB%B3%B4%EB%9D%BC%29.png",
         date = Date(),
         state = ExpenseState.NOT_CONFIRMED,
         categoryColor = "#FFFFFF"
