+++ ui/data/src/main/java/com/kappzzang/jeongsan/intentcontract/StartContract.kt
