+++ ui/expenselist/src/main/java/com/kappzzang/jeongsan/expenselist/inviteinfo/InviteInfoViewModel.kt
@@ -6,8 +6,8 @@ import com.kappzzang.jeongsan.model.MemberItem
 import com.kappzzang.jeongsan.usecase.GetInviteInfoUseCase
 import com.kappzzang.jeongsan.usecase.SendInviteMessageUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
-import kotlinx.coroutines.Dispatchers
 import javax.inject.Inject
+import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.launch
